# swagger.yml
openapi: 3.1.0
info:
  title: CatMapper API
  version: 1.1.0
  description: |
    Public and authenticated endpoints for CatMapper (SocioMap and ArchaMap).

    Canonical REST endpoints are now exposed under `/api`. Existing unprefixed
    routes remain available as deprecated compatibility aliases during the
    migration.

    Notes on auth:
    - Most read endpoints do not require authentication.
    - Write/task endpoints support either:
      1) `Authorization: Bearer <token>`
      2) `X-API-Key: <key>` (or `Authorization: ApiKey <key>`)
      3) Backward-compatible body credentials in `cred` (`{ userid, key }`).
servers:
  - url: https://api.catmapper.org/api
    description: Production canonical REST API
  - url: http://localhost:5000/api
    description: Local development canonical REST API
...
paths:
...
  /uploads/input-nodes:
    post:
      tags: [Upload]
      summary: Queue an input-node upload
      description: 'Canonical REST endpoint. Deprecated compatibility alias: `/uploadInputNodes`.'
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '202':
          description: Upload task queued
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /uploadInputNodes:
    post:
      tags: [Upload]
      summary: Start upload task
      description: |
        Queues upload processing and returns a task id.
        Mode matrix:
        - `standard`: caller supplies final `Key` expressions (for example `Field == Value`).
        - `simple`: only valid for `ao=add_uses`; caller supplies raw key values and server derives `Key`.
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadStartRequest'
      responses:
        '202':
          description: Upload task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadTaskResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /uploadInputNodesStatus:
    post:
      tags: [Upload]
      summary: Get upload task status
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusRequest'
      responses:
        '200':
          description: Upload task status
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /uploadInputNodesCancel:
    post:
      tags: [Upload]
      summary: Cancel upload task
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusRequest'
      responses:
        '202':
          description: Cancel request accepted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /uploadWaitingUSESStatus:
    post:
      tags: [Upload]
      summary: Get waiting-USES task status
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusRequest'
      responses:
        '200':
          description: Waiting-USES task status
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_details:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
                nullable: true
              field:
                type: string
              code:
                type: string
              message:
                type: string
            additionalProperties: true
      additionalProperties: true
    LegacyCred:
      type: object
      properties:
        userid:
          type: string
        key:
          type: string
      additionalProperties: true
...
    UploadStartRequest:
      type: object
      required: [database, formData, df]
      properties:
        database:
          type: string
        user:
          type: string
        cred:
          $ref: '#/components/schemas/LegacyCred'
        so:
          type: string
          description: Upload mode selector (`standard` or `simple`). `simple` is only valid with `ao=add_uses` and raw (non-`==`) key values.
        ao:
          type: string
          description: Upload action option (for example `add_uses`).
        formData:
          type: object
          additionalProperties: true
        df:
          type: array
          items:
            type: object
            additionalProperties: true
        addoptions:
          type: object
          additionalProperties: true
        optionalProperties:
          type: array
          description: Preferred list of editable/upload property columns. `variable` is treated as a plain USES relationship property.
          items:
            type: string
        allContext:
          type: array
          description: Deprecated alias for `optionalProperties`.
          items:
            type: string
        ignore_if_same:
          type: boolean
          description: If true, `update_add` skips rows where an existing string property already equals the incoming value.
        cursor:
          type: integer
      additionalProperties: true
    TaskStatusRequest:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: string
        user:
          type: string
        cred:
          $ref: '#/components/schemas/LegacyCred'
        cursor:
          type: integer
      additionalProperties: true
    UploadTaskResponse:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
          example: queued
        queue:
          type: string
          example: rq
        progress:
          type: object
          properties:
            batchSize:
              type: integer
            totalRows:
              type: integer
            totalBatches:
              type: integer
            completedBatches:
              type: integer
            percent:
              type: integer
          additionalProperties: true
        warnings:
          type: array
          items:
            type: string
      additionalProperties: true